refactored oauth routes, roomscontroller
